refactoring console debugger class, took out the common functionality
