+++ src/main/kotlin/org/web3j/evm/entity/source/SourceMapElement.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Web3 Labs Ltd.
+ * Copyright 2022 Web3 Labs Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -10,15 +10,6 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.web3j.evm
+package org.web3j.evm.entity.source
 
-import java.net.URL
-import org.web3j.abi.datatypes.Address
-
-class Configuration @JvmOverloads constructor(
-    selfAddress: Address,
-    val ethFund: Long,
-    val genesisFileUrl: URL? = null
-) {
-    val testAddress: String = selfAddress.value
-}
+data class SourceMapElement(val sourceFileByteOffset: Int = 0, val lengthOfSourceRange: Int = 0, val sourceIndex: Int = 0, val jumpType: String = "")
